// msfs-avionics-mirror/src/sdk/flightplan/FlightPlanning.ts
export type VectorTurnDirection = 'left' | 'right';

/**
 * The transition type to which a flight path vector belongs.
 */
export enum FlightPathVectorFlags {
  None,
  TurnToCourse = 1 << 0,
  Arc = 1 << 1,
  HoldInboundLeg = 1 << 2,
  HoldOutboundLeg = 1 << 3,
  HoldDirectEntry = 1 << 4,
  HoldTeardropEntry = 1 << 5,
  HoldParallelEntry = 1 << 6,
  CourseReversal = 1 << 7,
  LegToLegTurn = 1 << 8,
  AnticipatedTurn = 1 << 9,
}

/**
 * A basic flight path vector.
 */
export interface BaseFlightPathVector {
  /** The type of vector. */
  vectorType: string;

  /** Bit flags describing the vector. */
  flags: number;

  /** The latitude of the start of the vector. */
  startLat: number;

  /** The longitude of the start of the vector. */
  startLon: number;

  /** The latitude of the end of the vector. */
  endLat: number;

  /** The longitude of the end of the vector. */
  endLon: number;

  /** The total distance of the vector, in meters. */
  distance: number;
}

/**
 * A flight path vector whose path is defined by a geo circle.
 */
export interface CircleVector extends BaseFlightPathVector {
  /** The type of vector. */
  vectorType: 'circle';

  /** The radius of the circle, in great-arc radians. */
  radius: number;

  /** The x-coordinate of the center of the circle. */
  centerX: number;

  /** The y-coordinate of the center of the circle. */
  centerY: number;

  /** The z-coordinate of the center of the circle. */
  centerZ: number;
}

/**
 * A flight path vector within a leg flight path calculation.
 */
export type FlightPathVector = CircleVector;
// ...
export interface LegCalculations {

  /** The initial DTK of the leg. */
  initialDtk: number | undefined;

  /** The leg's total distance in meters, not cut short by ingress/egress turn radii. */
  distance: number;

  /** The cumulative distance in meters up to this point in the flight plan. */
  cumulativeDistance: number;

  /** The leg's total distance in meters, with leg transition turns take into account. */
  distanceWithTransitions: number;

  /** The cumulative distance in meters up to this point, with leg transition turns taken into account. */
  cumulativeDistanceWithTransitions: number;

  /** The latitude of the start of the leg. */
  startLat: number | undefined;

  /** The longitude of the start of the leg. */
  startLon: number | undefined;

  /** The latitude of the end of the leg. */
  endLat: number | undefined;

  /** The longitude of the end of the leg. */
  endLon: number | undefined;

  /** The calculated flight path for the leg. */
  flightPath: FlightPathVector[];

  /** The leg's flight path ingress transition. */
  ingress: FlightPathVector[];

  /** The index of the flight path vector in `flightPath` to which the ingress transition is joined. */
  ingressJoinIndex: number;

  /** The leg's flight path between the ingress and egress transitions. */
  ingressToEgress: FlightPathVector[];

  /** The index of the flight path vector in `flightPath` to which the egress transition is joined. */
  egressJoinIndex: number;

  /** The leg's flight path egress transition. */
  egress: FlightPathVector[];
}